# cmd/gomodel/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "High-performance AI gateway routing requests to multiple LLM providers (OpenAI, Anthropic, Gemini, Groq, xAI, Oracle, Ollama). Drop-in OpenAI-compatible API.",
        "title": "GOModel API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/v1/chat/completions": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json",
                    "text/event-stream"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "Create a chat completion",
                "parameters": [
                    {
                        "description": "Chat completion request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/core.ChatRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "JSON response or SSE stream when stream=true",
                        "schema": {
                            "$ref": "#/definitions/core.ChatResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/core.GatewayError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/core.GatewayError"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/core.GatewayError"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/core.GatewayError"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
    "definitions": {
...
        "core.ChatRequest": {
            "type": "object",
            "properties": {
                "max_tokens": {
                    "type": "integer"
                },
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/core.Message"
                    }
                },
                "model": {
                    "type": "string"
                },
                "parallel_tool_calls": {
                    "type": "boolean"
                },
                "provider": {
                    "description": "Gateway routing hint; stripped before upstream execution.",
                    "type": "string"
                },
                "reasoning": {
                    "$ref": "#/definitions/core.Reasoning"
                },
                "stream": {
                    "type": "boolean"
                },
                "stream_options": {
                    "$ref": "#/definitions/core.StreamOptions"
                },
                "temperature": {
                    "type": "number"
                },
                "tool_choice": {
                    "description": "string or object"
                },
                "tools": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "additionalProperties": {}
                    }
                }
            }
        },
        "core.ChatResponse": {
            "type": "object",
            "properties": {
                "choices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/core.Choice"
                    }
                },
                "created": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "object": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                },
                "system_fingerprint": {
                    "type": "string"
                },
                "usage": {
                    "$ref": "#/definitions/core.Usage"
                }
            }
        },
        "core.Choice": {
            "type": "object",
            "properties": {
                "finish_reason": {
                    "type": "string"
                },
                "index": {
                    "type": "integer"
                },
                "logprobs": {
                    "type": "object",
                    "properties": {
                        "content": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "properties": {
                                    "bytes": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    },
                                    "logprob": {
                                        "type": "number"
                                    },
                                    "token": {
                                        "type": "string"
                                    },
                                    "top_logprobs": {
                                        "type": "array",
                                        "items": {
                                            "type": "object",
                                            "properties": {
                                                "bytes": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "integer"
                                                    }
                                                },
                                                "logprob": {
                                                    "type": "number"
                                                },
                                                "token": {
                                                    "type": "string"
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "message": {
                    "$ref": "#/definitions/core.ResponseMessage"
                }
            }
        },
        "core.CompletionTokensDetails": {
            "type": "object",
            "properties": {
                "accepted_prediction_tokens": {
                    "type": "integer"
                },
                "audio_tokens": {
                    "type": "integer"
                },
                "reasoning_tokens": {
                    "type": "integer"
                },
                "rejected_prediction_tokens": {
                    "type": "integer"
                }
            }
        },
        "core.ContentPart": {
            "type": "object",
            "properties": {
                "image_url": {
                    "$ref": "#/definitions/core.ImageURLContent"
                },
                "input_audio": {
                    "$ref": "#/definitions/core.InputAudioContent"
                },
                "text": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "core.ErrorType": {
            "type": "string",
            "enum": [
                "provider_error",
                "rate_limit_error",
                "invalid_request_error",
                "authentication_error",
                "not_found_error"
            ],
            "x-enum-varnames": [
                "ErrorTypeProvider",
                "ErrorTypeRateLimit",
                "ErrorTypeInvalidRequest",
                "ErrorTypeAuthentication",
                "ErrorTypeNotFound"
            ]
        },
...
        "core.FunctionCall": {
            "type": "object",
            "properties": {
                "arguments": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "core.GatewayError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "x-nullable": true
                },
                "message": {
                    "type": "string"
                },
                "param": {
                    "type": "string",
                    "x-nullable": true
                },
                "provider": {
                    "type": "string"
                },
                "status_code": {
                    "type": "integer"
                },
                "type": {
                    "$ref": "#/definitions/core.ErrorType"
                }
            }
        },
        "core.ImageURLContent": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string"
                },
                "media_type": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "core.InputAudioContent": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string"
                },
                "format": {
                    "type": "string"
                }
            }
        },
        "core.Message": {
            "type": "object",
            "properties": {
                "content": {
                    "description": "ContentSchema documents that `content` accepts either a plain string\nor an array of ContentPart values.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/core.ContentPart"
                    },
                    "x-oneof": "[{\"type\":\"null\"},{\"type\":\"string\"},{\"type\":\"array\",\"items\":{\"$ref\":\"#/definitions/core.ContentPart\"}}]"
                },
                "role": {
                    "type": "string"
                },
                "tool_call_id": {
                    "type": "string"
                },
                "tool_calls": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/core.ToolCall"
                    }
                }
            }
        },
...
        "core.PromptTokensDetails": {
            "type": "object",
            "properties": {
                "audio_tokens": {
                    "type": "integer"
                },
                "cached_tokens": {
                    "type": "integer"
                },
                "image_tokens": {
                    "type": "integer"
                },
                "text_tokens": {
                    "type": "integer"
                }
            }
        },
        "core.Reasoning": {
            "type": "object",
            "properties": {
                "effort": {
                    "description": "Effort controls how much reasoning effort the model should use.\nValid values are \"low\", \"medium\", and \"high\".",
                    "type": "string"
                }
            }
        },
        "core.ResponseMessage": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/core.ContentPart"
                    },
                    "x-oneof": "[{\"type\":\"null\"},{\"type\":\"string\"},{\"type\":\"array\",\"items\":{\"$ref\":\"#/definitions/core.ContentPart\"}}]"
                },
                "role": {
                    "type": "string"
                },
                "tool_calls": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/core.ToolCall"
                    }
                }
            }
        },
...
        "core.StreamOptions": {
            "type": "object",
            "properties": {
                "include_usage": {
                    "description": "IncludeUsage requests token usage information in streaming responses.\nWhen true, the final streaming chunk will include usage statistics.",
                    "type": "boolean"
                }
            }
        },
        "core.ToolCall": {
            "type": "object",
            "properties": {
                "function": {
                    "$ref": "#/definitions/core.FunctionCall"
                },
                "id": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "core.Usage": {
            "type": "object",
            "properties": {
                "completion_tokens": {
                    "type": "integer"
                },
                "completion_tokens_details": {
                    "$ref": "#/definitions/core.CompletionTokensDetails"
                },
                "prompt_tokens": {
                    "type": "integer"
                },
                "prompt_tokens_details": {
                    "$ref": "#/definitions/core.PromptTokensDetails"
                },
                "raw_usage": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "total_tokens": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}